+++ 13-画画之手/72-对称随机游走八对称.py
@@ -21,8 +21,8 @@ last = (width // 2, height // 2)
 
 def get_background(x, y):
     global last
-    xx = np.round(np.linspace(last[0], x, 100)).astype(np.int)
-    yy = np.round(np.linspace(last[1], y, 100)).astype(np.int)
+    xx = np.round(np.linspace(last[0], x, 100)).astype(np.int32)
+    yy = np.round(np.linspace(last[1], y, 100)).astype(np.int32)
     cx, cy = width // 2 - 1, height // 2 - 1
     xx[xx >= cx] = (cx * 2 - xx)[xx >= cx]
     yy[yy >= cy] = (cy * 2 - yy)[yy >= cy]
